+++ packages/plugin-client/src/generators/groupedClientGenerator.tsx
@@ -1,5 +1,5 @@
 import { camelCase } from '@kubb/core/transformers'
-import type * as KubbFile from '@kubb/fs/types'
+import type { KubbFile } from '@kubb/core/fs'
 
 import { createReactGenerator } from '@kubb/plugin-oas'
 import { useOas, useOperationManager } from '@kubb/plugin-oas/hooks'