+++ packages/plugin-oas/src/SchemaMapper.ts
@@ -1,4 +1,4 @@
-import type * as KubbFile from '@kubb/fs/types'
+import type { KubbFile } from '@kubb/core/fs'
 
 export type SchemaKeywordMapper = {
   object: {
@@ -21,7 +21,7 @@ export type SchemaKeywordMapper = {
   date: { keyword: 'date'; args: { type?: 'date' | 'string' } }
   time: { keyword: 'time'; args: { type?: 'date' | 'string' } }
   datetime: { keyword: 'datetime'; args: { offset?: boolean; local?: boolean } }
-  tuple: { keyword: 'tuple'; args: { items: Schema[]; min?: number; max?: number, rest?: Schema } }
+  tuple: { keyword: 'tuple'; args: { items: Schema[]; min?: number; max?: number; rest?: Schema } }
   array: {
     keyword: 'array'
     args: { items: Schema[]; min?: number; max?: number; unique?: boolean }