+++ packages/plugin-oas/src/hooks/useOperationManager.ts
@@ -4,7 +4,7 @@ import { Oas } from '../components/Oas.tsx'
 
 import type { FileMetaBase, Plugin, ResolveNameParams } from '@kubb/core'
 
-import type * as KubbFile from '@kubb/fs/types'
+import type { KubbFile } from '@kubb/core/fs'
 import type { Operation, Operation as OperationType } from '@kubb/oas'
 import type { OperationSchemas } from '../types.ts'